milestone-1/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.117.1
h11==0.16.0
httptools==0.6.4
idna==3.10
mysql-connector-python==9.4.0
pydantic==2.11.9
pydantic-core==2.33.2
python-dotenv==1.1.1
pyyaml==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing-extensions==4.15.0
typing-inspection==0.4.1
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
